+++ src/theme/defaultColors.tsx
@@ -1,5 +1,85 @@
-import { RawColors, RawColor } from './theme.types';
+const colorScales = {
+  // GRAY
+  white: '#FFFFFF',
+  gray100: '#F5F5F5',
+  gray200: '#EEEEEE',
+  gray300: '#E0E0E0',
+  gray400: '#BDBDBD',
+  gray500: '#9E9E9E',
+  gray600: '#6E6E6E',
+  gray700: '#424242',
+  gray800: '#212121',
+  gray900: '#121212',
+  black: '#000000',
 
+  // BLUE
+  blue100: '#EAF0FF',
+  blue200: '#D6E1FF',
+  blue300: '#ADC4FF',
+  blue400: '#6690FF',
+  blue450: '#00C8F5',
+  blue500: '#336BFF',
+  blue600: '#2650BF',
+  blue700: '#1F4099',
+  blue800: '#142B66',
+  blue900: '#0A1533',
+
+  // PINK
+  pink100: '#F6E7EE',
+  pink200: '#FFD5E6',
+  pink300: '#FFAACD',
+  pink400: '#FF6AA8',
+  pink500: '#FF2B83',
+  pink600: '#CC2269',
+  pink700: '#AC135A',
+  pink800: '#78013A',
+  pink900: '#590F2E',
+
+  // GREEN
+  green100: '#EDFAE8',
+  green200: '#DCF5D2',
+  green300: '#B9EAA6',
+  green400: '#85DA62',
+  green450: '#D2F500',
+  green500: '#51CB20',
+  green600: '#41A21A',
+  green700: '#317A13',
+  green800: '#20510D',
+  green900: '#102906',
+
+  // TEAL
+  teal100: '#E5FDFC',
+  teal200: '#CCFCF9',
+  teal300: '#99F9F3',
+  teal400: '#00F0E1',
+  teal500: '#00C0B4',
+  teal600: '#009087',
+  teal700: '#00605A',
+  teal800: '#004844',
+  teal900: '#00302D',
+
+  // YELLOW
+  yellow100: '#FFFAE5',
+  yellow200: '#FFF3BD',
+  yellow300: '#FFE98A',
+  yellow400: '#FFCF00',
+  yellow500: '#E6BB00',
+  yellow600: '#C49F00',
+  yellow700: '#9E8000',
+  yellow800: '#614F00',
+  yellow900: '#332900',
+
+  // RED
+  red100: '#FFECEA',
+  red200: '#FFC7C1',
+  red300: '#FFA399',
+  red400: '#FF7E70',
+  red500: '#FF4733',
+  red600: '#CC3929',
+  red700: '#992B1F',
+  red800: '#661C14',
+  red900: '#330E0A',
+};
 const defaultColors = {
   // BRAND
   brandBlue: '#00C8F5',
@@ -32,19 +112,6 @@ const defaultColors = {
   gray7: '#F5F5F5', // Included in Dark Mode
   white: '#FFFFFF', // Included in Dark Mode
 
-  // GREEN SCALE
-  // Introduced for new Shareville component from the new UI designs
-  // Names are subject to change when new colors are finalized
-  green600: '#41A21A',
-  green200: '#DCF5D2',
-  green100: '#EDFAE8',
-
-  // BLUE SCALE
-  blue100: '#E5ECFF',
-  blue400: '#6690FF',
-  blue500: '#336BFF',
-  blue800: '#142B66',
-
   // FUNCTIONAL COLOURS FOR PRODUCT DEVELOPMENT
   cta: '#0046FF',
   ctaHover: '#003BD9',
@@ -56,7 +123,9 @@ const defaultColors = {
   index: '#FFCF00', // Included in Dark Mode
   sfdrArticle9: '#41A21A',
   sfdrArticle8: '#85DA62',
-} as RawColor;
+  graphPurple: '#9C88FF',
+  graphOrange: '#F0932B',
+};
 
 const grayPalette = [
   defaultColors.gray6,
@@ -68,8 +137,9 @@ const grayPalette = [
 
 export default {
   ...defaultColors,
+  ...colorScales,
   palettes: {
-    gray: [grayPalette, defaultColors.gray100, defaultColors.gray900, defaultColors.gray800],
+    gray: grayPalette,
     pink: [
       defaultColors.brandPink,
       defaultColors.complementaryPink1,
@@ -84,10 +154,10 @@ export default {
       defaultColors.complementaryBlue2,
       defaultColors.brandTurquoise,
       defaultColors.complementaryTurquoise1,
-      defaultColors.blue100,
-      defaultColors.blue400,
-      defaultColors.blue500,
-      defaultColors.blue800,
+      colorScales.blue100,
+      colorScales.blue400,
+      colorScales.blue500,
+      colorScales.blue800,
       ...grayPalette,
     ],
     green: [
@@ -107,4 +177,4 @@ export default {
       ...grayPalette,
     ],
   },
-} as RawColors;
+};